c2viewer_api/requirements.txt
aioredis==1.3.1
asgiref==3.2.3
astroid==2.3.1
async-timeout==3.0.1
asyncio==3.4.3
attrs==19.3.0
autobahn==20.3.1
Automat==20.2.0
cffi==1.14.0
channels==2.4.0
channels-redis==2.4.2
colorama==0.4.1
constantly==15.1.0
cryptography==2.8
cycler==0.10.0
daphne==2.4.1
Django==3.0.3
django-cors-headers==3.2.1
django-environ==0.4.5
django-rest-auth==0.9.5
django-templated-mail==1.1.1
djangorestframework==3.11.0
djoser==2.0.3
et-xmlfile==1.0.1
hiredis==1.0.1
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
isort==4.3.21
jdcal==1.4.1
joblib==0.14.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.2
matplotlib==3.1.1
mccabe==0.6.1
msgpack==0.6.2
mysqlclient==1.4.6
numpy==1.17.2
openpyxl==3.0.0
pandas==0.25.1
psycopg2==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pylint==2.4.2
pyOpenSSL==19.1.0
pyparsing==2.4.2
python-dateutil==2.8.0
python-dotenv==0.11.0
pytz==2019.2
scikit-learn==0.21.3
scipy==1.3.1
service-identity==18.1.0
six==1.12.0
skfeature-chappers==1.0.3
sqlparse==0.3.0
tornado==6.0.3
Twisted==20.3.0
txaio==20.1.1
typed-ast==1.4.0
virtualenv==16.7.7
virtualenvwrapper-win==1.2.5
websockets==8.1
wrapt==1.11.2
xlrd==1.2.0
zope.interface==5.0.0
